app/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pillow==12.1.0
pillow-heif==1.1.1
python-multipart==0.0.21
apscheduler==3.11.2
tzlocal==5.3.1
